Fix date comparison with episodes from previous year
In order to classify the podcast episodes depending on
their date, the logic was relying on year comparison for
each date. That worked well for episodes published
during the current year, but didn't produce any match if
the episode was published the previous year. That meant
that an episode published on December 31st would not be
in 'Yesterday' but in 'Older' when accessing the app on
January 1st.

This commit refactors the date comparison logic so it
doesn't rely on dates being in the same year (except for
'This Month').

https://gitlab.gnome.org/World/podcasts/-/issues/165