Add time-based caching mechanism for podcast cover art
Previously, cached podcast cover art was never invalidated.
When podcasts updated cover art, it would not be updated locally.
To correct this, a time-based caching mechanism has been introduced.
This invalidates cached images after a specific amount of time.
This has been set to 4 weeks.
Once a podcast cover has been stale for 1 month, it is re-downloaded.
To get updates earlier, a hash-based mechanism will be implemented.
The time-based mechanism is mostly a fallback to guarantee updates.

One column is added to the shows table, the timestamp image_cached.
A database migration has been included as part of this commit.

Testing has yet to be implemented for this functionality.