episode_description: added new widget
It was not possible to see the description text of episodes¹, or their
full title², or copy their file url for sharing.

This adds a new episode_description widget that is opened in a deck
above the content stack when activating an episode_widget.
Notable changes to app.rs and window.rs:
* added a new hdy::Deck main_deck above content to display pages like
the episode_view without the content stack bar.
* deck pages can have custom header bars that are placed above the
default header bar.
* replaced Action::InitShowMenu with more generic Action that just
initalizes any gio::MenuModel.

1) https://gitlab.gnome.org/World/podcasts/-/issues/93
2) https://gitlab.gnome.org/World/podcasts/-/issues/203