Store the image hash as blob
The hash of the image uri was previously stored as an integer in the db.
This has been changed to avoid casting in the hashing function.
The hash value, a u64, is now stored as 8 u8 values in a binary blob.
These 8 values are stored in a Vec<u8> instead of an array for the db.
Conversion functions have been added as required.

Furthermore, a bug involving hashing the image uri has been fixed.
Previously, the hashing function was erroneously hashing the Option.
Now, the hashing function only hashes the String value, as expected.