Revert "podcasts-gtk: Remove static resources"
This reverts commit f8426e364f261a5af4515ab2eff4c7d0463e9987.

Installing it on the path and opening it is not great, given
that the gresources are part of the app and can be linked against
the binary. The only advantage is not having to relink the
application if you only changed resource files, which isn't a
usecase that needs to be optimized for.

Additionally this avoids opening and loading a couple megabytes
of assets from the filesystem during startup, which is minor
but does make startup a bit faster.

This also avoids weird issues like #301 and the tests are now
working again even without installing the application first.

Close #301