episode_description: add stream/download/play/delete buttons.
Not being able to play an episode from the description page was a huge
usability and accessibility pain point. Also the home view didn't have
enough space for a streaming button. Both these issues should be fixed
here.

- I separated the download_progress_bar from the episode widget into a
separate widget.
- I added an option to call the player with a stream option
that will pass the http/s url to gst instead of the local file url.
- I added buttons for streaming/downloading/playing/deleting downloads
to the episode description page.