lazy_load: Improve the loading of widgets
Previously we were creating a widget and scheduling an idle_add
per widget we wanted to add/insert into the container.

This causes problems, as we often invalidate the layout and
so each insert ends up queing up a revalidation of the CSS.

Start inserting widgets in batches with a timed callback
instead of idle_add so we avoid ddosing and blocking the
main thread for long.