gtk: Create the widgets in the background before inserting them
Previously we were creating the widgets and inserting them at the
same time. While we want to batch the insertions to avoid triggering
multiple stylesheets invalidations, we can create all the widgets
in the background at once, and then only start batchign the insertions.

This way we reduce significantly the time we block the main thread
when creating widgets, as we can yield after each one, while maintaining
the batching of widget insertions into the containers.