lazy_load: Replace insert widget callback
This was now only used to add widgets to the container. Move the
code inline, in the idle_add callack. In order to do this, since
the append(or other method in the future) all come from different
Traits, we had to replace Rust's generics with trying to cast
the widget at runtime, and then calling the appropriate method.