widgets/episode: defer initialization of the widget
Instead of initializing the widget when we create it, defer that
to a future that can be scheduled later.

This allows us to defer 50-100ms per widget created.

This doesn't yet do anything to distribute the work or yield back,
everything is still done syncronously, but we can start using .await
now during the widget init.