data: Don't identify episodes by their title when they have a guid
Removes the (title,show_id) primary key from episodes.
Uniqueness is now checked by a query that will first check for (guid,show_id).
If no guid is set, the old behavior of checking (title,show_id) is used.

fixes #151 #204 #215 #216 #239 #317

DB changes in the episodes table:
---
rowid was renamed to id to be consistent with other tables. Also
sqlite will by default use it as an alias to rowid, so there should be
no downsides there.

An Id is used instead of adding guid to a combined primary key
of (guid,title,show_id), because two episodes with the same non-NULL
guid are considered the same even with the a different title.

In the future it could be considered to move to a hash-id field that
is generated from (guid,show_id)/(title,show_id)/... this could
prevent double entries at db-level, but is more complicated to implement.

With this implementation double entries will be prevented in
NewEpisode::index() which is already used over NewEpisode::insert().
Also queries that only check for title uniqueness have been made private.

Other bug fixes:
---
A bug in the batch inserter was also fixed where a feed containing
multiples of the same episode, was not filtered for duplicates and
always failed to import.