window: pop pages before going to the Show widget page.
It was possible to construct chains of multiple
episode_description pages by navigating like this by using the
"Go to Show" menu point on episode_description:

Home -> Episode -> Show -> Episode -> Show -> Episode

The show page would be poped, but the episode pages would remain until
the back button was used.

Also Action::ReplaceWidget would auto navigate to the Show page, when it
should only update it. Action::GoToShow should be used to navigate to
the Show page instead.

Now the navigation pops back to an existing show page when trying to
go to the Show page.