feed_manager: add a locked feed refresh manager
Right now only one update can run and it's not possible to add new
feed updates while it still runs. It is also not possible to properly
await the end of the update.

With this feed manager updates can be started for new feeds while the
old feeds are still updating. And updates can be awaited in async.

This should fix a possible conflict when adding a feed via the
discovery_page and also improve the loading spinner behavior by having
it be visible until the feed is actually done loading.