chapters: add chapter parser and UI for listing chapters.
Previously chapters in media files would not be utilized and only jump
links in the description could be used to navigate to points in an episode.

Now when an episode gets played this patch adds a buttons in the big
player controls, the player sheet. The button opens a new Navigation
page on which the chapters for the currently playing Episode are
listed. Clicking/pressing the chapters jumps to them.