download_covers: update cache timestamp after download.
Previously covers would not update their timestamp in the DB, which
would cause them to be refetched too often on older databases.

This updates the timestamp and hash for the URL.

All cache update functions are moved to ShowCoverModel from Show, since
that is the only place they are ever used.