chrono: enable localization.
Currently month names always display in English.

This patch uses the chrono unstable-locales feature to get the
format_localized method. It is unstable, because the API is
experimental and may be removed in the future.

The current locale is retrieved with the locale_config crate which is
already used by gettextrs.