diesel: store epoch and played as TIMESTAMP instead of integer.
Previously epoch and played dates were i32 values in the Episode model
struct. This would have them overflow in 2038.

This changes the i32 values to NaiveDateTime.
Using diesel's default NaiveDateTime compatability requires the dates
to be stored as TEXT/TIMESTAMP instead of INTEGER, so I added a
migration to convert them.