thumbnail_generator: guess the image format.
Some images are named png and send as png mime type from
the server, but are actually jpegs, in that case the thumbnail
generator would fail to load the image.

This patch uses image-rs's format guesser to determine the image type.
With this images load even if they are named wrong.