gtk: use result API of html2text.
Currently html2text::from_read panics when width is 0. So Podcasts
with an empty description crash the application when viewing their
show page.

This patch switches to the newer config-API, which will return a
Result and not panic. If html2text fails, the description is displayed
raw and unmodified.

Fixes #351