episode_description: fix accessibility warnings.
The new icons in buttons were not marked as presentation.
Their meaning is already described by the button.

The Episode Cover image was not described previously.
There doesn't seem to be alt text for it in the rss feeds.
So I just labeled it as Episode Cover to genericly clarify where it
comes from and what it shows.