window: fix both the bottom and top switcher being visible at once.
The visiblity of the bottom switcher is determined by both the window
size/layout and on what page you're on. Currently both directly set
the visibilty which can lead to situations where the switcher becomes
visible even though only one condition for it's visibilty is
justified.

Resizing the window and going to an episode page and back could both
lead to the bottom switcher being visible when it shouldn't be.

With this patch layout-dependent-visibilty and
page-dependent-visibilty are split up into two properties.
And the bottom switcher is only set to visible when both of these are
true.