player: add a button to the sheet that opens the episode description
There is an expectation to see more about the current episode when
sliding up the bottom sheet of what's currently playing. Currently you
have to manually find the playing episode in the Home or show view
again. On phones where you might pause for multiple days this can
become especially difficult.

This patch adds a button that closes the bottom sheet and opens the
description page for the currently playing episode. It also prevents
stacking multiple episode widgets ontop of eachother.

Closes #365