new_episode: use rss content:encoded tag before description tag.
Some podcasts provide a better¹ description in the content:encoded tag,
or don't provide anything useful in the description tag at all.

For example see:
https://feeds.lagedernation.org/feeds/ldn-mp3.xml
which only has the episode title in the description tag, but long show
notes in content.

The Episode Description Parser can handle these and this will improve
the description page.

1) https://www.rssboard.org/rss-profile#namespace-elements-content-encoded