home/discovery: fix covers being small & centered, new missing-icon.
The problem:

As of gtk-4.19.2 GTK changed GtkImage behavior when displaying
GdkPaintable to strictly use the :pixel-size property property instead
of stretching the paintable to the allocated size.¹
This leads Podcasts to show covers on the home-page and on the
discovery-results-listing as small centered image inside a rectangle.

It can be easily fixed by changing the pixel-size to match the
width-request of the cover image, but the request is currently set to
half-size to make the image-missing icon centered.

I prefer to not add any code that manually changes the pixel size in
different cases just always keep it at 64px. The current codebase is
already too complicated with little workarounds like that.

The current missing icon is a white icon that gets faded-out by
lowering the opacity and given a background in css. It gets centered
by the lower pixel-width. Other pages like shows and description don't
bother with centering, because the Images there don't have a
background they can be centered on. Graying out has been
inconsistent between pages.

The fix:

By switching to `image-missing-symbolic` (as suggested by #289) from
`image-x-generic-symbolic` we get an icon that is already grayed-out
and has it's own dashed-border to delineate the Image, so we can
remove the background and ditch missing-icon centering.
To make it consistent I also switched all other missing images to that
icon.

Setting :pixel-size to the width/height-request fixes the centering
bug.

1) https://thisweek.gnome.org/posts/2025/06/twig-203/#gtk