episode: hide duration label when no duration is available.
Currently the default duration of "42 min" gets displayed for episodes
that either lack the duration in the RSS feed (all patreon feeds) or
have a low duration of < 1 min. This is wrong and a bug. This happens
because later state changes make the duration label visible again
without checking if it has a valid value.

With this patch the duration label gets set to an empty string when
there is no valid duration. Setting the visibility of the duration
label now checks if the label is non-empty and also manages the
visibility of separator1.

Comments about separator1 have been removed. separator1 belongs to the
duration label and should not be thought of separately.

I don't think this bug was reported previously.