description: allow horizontal scrollbar to avoid ui width overflow.
Currently there is some bug with specific descriptions not wrapping
and causing layout-width to become bigger than the window.

See: https://gitlab.gnome.org/World/podcasts/-/issues/398

This allows horizontal scrollbars for the description. It's not a good
solution, but keeps the ui usable in these cases and allows us to make
a release to fix other bugs and introduce the chapters bottom sheet.