sync: use rfc3339 dates for episode actions.
Thanks to Simon Kindhauser for finding and fixing this.
https://gitlab.gnome.org/SimonKindhauser

By Nextcloud 32 the GPodderSync plugin uses rfc dates.
The previous code would fail to parse those, expecting no timezone.

This patch falls back to using rfc3339 dates in that case.
The old date parsing is still kept to support older servers for now,
or in case it may come in useful in future support for GPodder.net servers.