Port to gettext with formatx macro
See https://www.gnu.org/software/gettext/manual/html_node/Rust.html.

This requires xgettext 0.24, otherwise the translatable strings won't be
extracted. As of now d-l uses version 0.25.

When running `meson compile gnome-podcasts-pot`, the format-able strings
are marked with `rust-format`.

It is worth mentioning that we were not extracting strings with plurals
previously and this fixes it.

The only strings that were modified where the "Jump to {}:{}" and "Jump
to {}:{}:{}" for ease of translating.